DDCNL-9021 | moved /eligible-years into its own controller &  tests
