+++ app/controllers/TransferController.scala
@@ -109,63 +109,63 @@ class TransferController @Inject() (
 //      }
 //    ) recover handleError
 //  }
-
-  def eligibleYears: Action[AnyContent] = authenticate.pertaxAuthActionWithUserDetails.async { implicit request =>
-    registrationService.deleteSelectionAndGetCurrentAndPreviousYearsEligibility map {
-      case CurrentAndPreviousYearsEligibility(false, Nil, _, _) =>
-        throw new NoTaxYearsAvailable
-      case CurrentAndPreviousYearsEligibility(false, previousYears, registrationInput, _) if previousYears.nonEmpty =>
-        Ok(previousYearsV(registrationInput, previousYears, currentYearAvailable = false))
-      case CurrentAndPreviousYearsEligibility(_, previousYears, registrationInput, _) =>
-        Ok(
-          eligibleYearsV(
-            currentYearForm(previousYears.nonEmpty),
-            previousYears.nonEmpty,
-            registrationInput.name,
-            Some(registrationInput.dateOfMarriage),
-            Some(timeService.getStartDateForTaxYear(timeService.getCurrentTaxYear))
-          )
-        )
-    } recover handleError
-  }
-
-  def eligibleYearsAction: Action[AnyContent] = authenticate.pertaxAuthActionWithUserDetails.async { implicit request =>
-    registrationService.getCurrentAndPreviousYearsEligibility.flatMap {
-      case CurrentAndPreviousYearsEligibility(currentYearAvailable, previousYears, registrationInput, _) =>
-        currentYearForm(previousYears.nonEmpty).bindFromRequest().fold(
-          hasErrors =>
-            Future {
-              BadRequest(
-                eligibleYearsV(
-                  hasErrors,
-                  previousYears.nonEmpty,
-                  registrationInput.name,
-                  Some(registrationInput.dateOfMarriage),
-                  Some(timeService.getStartDateForTaxYear(timeService.getCurrentTaxYear))
-                )
-              )
-            },
-          success => {
-
-            val selectedYears = if (success.applyForCurrentYear.contains(true)) {
-              List[Int](timeService.getCurrentTaxYear)
-            } else {
-              List[Int]()
-            }
-
-            registrationService.saveSelectedYears(selectedYears) map { _ =>
-              if (previousYears.isEmpty && currentYearAvailable && (!success.applyForCurrentYear.contains(true))) {
-                throw new NoTaxYearsSelected
-              } else if (previousYears.nonEmpty) {
-                Ok(previousYearsV(registrationInput, previousYears, currentYearAvailable))
-              } else {
-                Redirect(controllers.routes.TransferController.confirmYourEmail())
-              }
-            }
-          }
-        )
-    } recover handleError
-  }
+//
+//  def eligibleYears: Action[AnyContent] = authenticate.pertaxAuthActionWithUserDetails.async { implicit request =>
+//    registrationService.deleteSelectionAndGetCurrentAndPreviousYearsEligibility map {
+//      case CurrentAndPreviousYearsEligibility(false, Nil, _, _) =>
+//        throw new NoTaxYearsAvailable
+//      case CurrentAndPreviousYearsEligibility(false, previousYears, registrationInput, _) if previousYears.nonEmpty =>
+//        Ok(previousYearsV(registrationInput, previousYears, currentYearAvailable = false))
+//      case CurrentAndPreviousYearsEligibility(_, previousYears, registrationInput, _) =>
+//        Ok(
+//          eligibleYearsV(
+//            currentYearForm(previousYears.nonEmpty),
+//            previousYears.nonEmpty,
+//            registrationInput.name,
+//            Some(registrationInput.dateOfMarriage),
+//            Some(timeService.getStartDateForTaxYear(timeService.getCurrentTaxYear))
+//          )
+//        )
+//    } recover handleError
+//  }
+//
+//  def eligibleYearsAction: Action[AnyContent] = authenticate.pertaxAuthActionWithUserDetails.async { implicit request =>
+//    registrationService.getCurrentAndPreviousYearsEligibility.flatMap {
+//      case CurrentAndPreviousYearsEligibility(currentYearAvailable, previousYears, registrationInput, _) =>
+//        currentYearForm(previousYears.nonEmpty).bindFromRequest().fold(
+//          hasErrors =>
+//            Future {
+//              BadRequest(
+//                eligibleYearsV(
+//                  hasErrors,
+//                  previousYears.nonEmpty,
+//                  registrationInput.name,
+//                  Some(registrationInput.dateOfMarriage),
+//                  Some(timeService.getStartDateForTaxYear(timeService.getCurrentTaxYear))
+//                )
+//              )
+//            },
+//          success => {
+//
+//            val selectedYears = if (success.applyForCurrentYear.contains(true)) {
+//              List[Int](timeService.getCurrentTaxYear)
+//            } else {
+//              List[Int]()
+//            }
+//
+//            registrationService.saveSelectedYears(selectedYears) map { _ =>
+//              if (previousYears.isEmpty && currentYearAvailable && (!success.applyForCurrentYear.contains(true))) {
+//                throw new NoTaxYearsSelected
+//              } else if (previousYears.nonEmpty) {
+//                Ok(previousYearsV(registrationInput, previousYears, currentYearAvailable))
+//              } else {
+//                Redirect(controllers.routes.TransferController.confirmYourEmail())
+//              }
+//            }
+//          }
+//        )
+//    } recover handleError
+//  }
 
   def previousYears: Action[AnyContent] = authenticate.pertaxAuthActionWithUserDetails.async { implicit request =>
     registrationService.getCurrentAndPreviousYearsEligibility.map {
