+++ app/views/multiyear/transfer/email.scala.html
@@ -38,7 +38,7 @@ request: AuthenticatedUserRequest[_])
 @main(
     pageTitle = viewHelpers.formErrorPrefix(recipientForm, Messages("title.confirm-email")),
     serviceTitle = "title.application.pattern",
-    backLinkHref = Some(controllers.routes.TransferController.eligibleYears().url)
+    backLinkHref = Some(controllers.transfer.routes.EligibleYearsController.eligibleYears().url)
 ){
 
     @form(action = routes.TransferController.confirmYourEmailAction(), args = (Symbol("id") -> "register-form")) {
