+++ conf/app.routes
@@ -20,8 +20,8 @@ GET         /date-of-marriage                   controllers.transfer.DateOfMarri
 GET         /date-of-marriage-en                controllers.transfer.DateOfMarriageController.dateOfMarriageWithEn()
 GET         /date-of-marriage-cy                controllers.transfer.DateOfMarriageController.dateOfMarriageWithCy()
 POST        /date-of-marriage                   controllers.transfer.DateOfMarriageController.dateOfMarriageAction()
-GET         /eligible-years                     controllers.TransferController.eligibleYears()
-POST        /eligible-years                     controllers.TransferController.eligibleYearsAction()
+GET         /eligible-years                     controllers.transfer.EligibleYearsController.eligibleYears()
+POST        /eligible-years                     controllers.transfer.EligibleYearsController.eligibleYearsAction()
 GET         /previous-years                     controllers.TransferController.previousYears()
 POST        /extra-years                        controllers.TransferController.extraYearsAction()
 GET         /confirm-your-email                 controllers.TransferController.confirmYourEmail()
